task 2 (choose a date)
The onClose() method from the options object is called each time the UI element that flatpickr creates is closed. It is in it that the date selected by the user should be processed. The selectedDates parameter is an array of selected dates, so we take the first element.

If the user has selected a date in the past, show window.alert() with the text "Please choose a date in the future".
If the user has selected a valid date (in the future), the "Start" button becomes active.
The "Start" button should be disabled until the user has selected a date in the future.
When you click on the "Start" button, the countdown to the selected date starts from the moment you click it.