task 3
Write a script that, when submitting a form, calls the createPromise(position, delay) function as many times as you enter in the amount field. At each call, pass it the number of the generated promise (position) and the delay, taking into account the first delay (delay) and step (step) entered by the user.